Wsign-conversion warning
fixed a -Wsign-conversion warning in gcc and clang by static casting the iterator difference to a size_t. The result should should always be >=0 so the static cast will just suppress the warning

+++ single_include/clara.hpp
@@ -489,7 +489,7 @@ namespace detail {
             return !m_tokenBuffer.empty() || it != itEnd;
         }
 
-        auto count() const -> size_t { return m_tokenBuffer.size() + (itEnd - it); }
+        auto count() const -> size_t { return m_tokenBuffer.size() + static_cast<size_t>(itEnd - it); }
 
         auto operator*() const -> Token {
             assert( !m_tokenBuffer.empty() );
